requirements.txt
evalutils==0.3.1
fastai==1.0.61
hydra-core==1.3.2
numpy==1.21.6
object_detection_fastai==0.0.10
omegaconf==2.2.3
opencv_python_headless==4.5.5.64
openslide_python==1.1.2
pandas==1.1.2
Pillow==9.4.0
PyYAML==6.0
scikit_learn==1.0
SimpleITK==2.2.1
SlideRunner==2.0
torchmetrics==0.11.0
torchvision==0.8.1+cu101
tqdm==4.49.0
